Create codeql-analysis.yml
Creates code-ql file for "getip", the current name of the program that synchronizes the "outside ips" DHCP ipv4 address of two gateways (in this case, comcast.net) and writes these addresses into the "hosts" file of the target computers.

The target computers are identified in the "port forwarding" section of your gateway setup. The port numbers indicate particular services (ssh, ftp, etc.). 

The individual computers are identified as more names with the same address of the gateway, viz.
69.180.153.161    downstairs  computer1 computer2 computer3
, where "downstairs is the name of the comcast.net gateway in the downstairs apartment.